+++ tests/testthat/test-step-reduce.R
@@ -104,3 +104,20 @@ test_that("collect() calls `as_iterator()`", {
     list("foo")
   )
 })
+
+test_that("collect() preserves type (#32)", {
+  expect_equal(
+    collect(gen(for (i in 1:3) yield(i))),
+    as.list(1:3)
+  )
+
+  expect_equal(
+    collect(gen(for (i in 1:3) yield(list(i)))),
+    lapply(1:3, list)
+  )
+
+  expect_equal(
+    collect(gen(for (i in 1:3) yield(mtcars[i, 1:3]))),
+    lapply(1:3, function(i) mtcars[i, 1:3])
+  )
+})